library_system/requirements.txt
Flask==2.0.1
mysql-connector-python==9.1.0
Flask-Login==0.5.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
python-dotenv==0.19.0
matplotlib==3.7.1
pandas==2.0.0
